located HRS variables to replicate Guiso 2008
Ive created a .dta file which contains the variables necessary to replicate the Guiso et al. results regarding the relationship between trust and participation and shares. Note, the HRS does not provide a question/elicitation of risk/ambiguity aversion or optimism. In addition, I've left out the variables from the module "potential for deception", which may be of interest later.